Splits the base client and 2 implementations into separate files.
