+++ lib/twilio-ruby.rb
@@ -78,6 +78,7 @@ require 'twilio-ruby/rest/notifications'
 require 'twilio-ruby/rest/addresses'
 require 'twilio-ruby/rest/addresses/dependent_phone_numbers'
 require 'twilio-ruby/rest/client'
+require 'twilio-ruby/rest/task_router_client'
 require 'rack/twilio_webhook_authentication'
 
 module Twilio
